+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - raise on invalid UInt8 and Int8 when encoding RowBinary https://github.com/plausible/ch/pull/180
+- - move rowbinary from `params` to `statement`
 
 ## 0.2.6 (2024-05-30)